% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{topicsTest}
\alias{topicsTest}
\title{The function to test the lda model}
\usage{
topicsTest(
  model,
  preds,
  data,
  pred_var = NULL,
  group_var = NULL,
  control_vars = c(),
  test_method = "linear_regression",
  p_adjust_method = "fdr",
  seed = 42,
  load_dir = NULL,
  save_dir = "./results"
)
}
\arguments{
\item{model}{(list) The trained model}

\item{preds}{(tibble) The predictions}

\item{data}{(tibble) The data to test on}

\item{pred_var}{(string) The variable to be predicted (only needed for regression or correlation)}

\item{group_var}{(string) The variable to group by (only needed for t-test)}

\item{control_vars}{(vector) The control variables}

\item{test_method}{(string) The test method to use, either "correlation","t-test", "linear_regression","logistic_regression", or "ridge_regression"}

\item{p_adjust_method}{(character) Method to adjust/correct p-values for multiple comparisons
(default = "none"; see also "holm", "hochberg", "hommel", "bonferroni", "BH", "BY",  "fdr").}

\item{seed}{(integer) The seed to set for reproducibility}

\item{load_dir}{(string) The directory to load the test from, if NULL, the test will not be loaded}

\item{save_dir}{(string) The directory to save the test, if NULL, the test will not be saved}
}
\value{
A list of the test results, test method, and prediction variable
}
\description{
The function to test the lda model
}
