% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{topicsDtm}
\alias{topicsDtm}
\title{the function for creating a document term matrix}
\usage{
topicsDtm(
  data,
  id_col,
  data_col,
  ngram_window = c(1, 3),
  stopwords = stopwords::stopwords("en", source = "snowball"),
  removalword = "",
  occ_rate = 0,
  removal_mode = "",
  removal_rate_most = 0,
  removal_rate_least = 0,
  split = 1,
  seed = 42,
  save_dir = "./results"
)
}
\arguments{
\item{data}{(tibble) the data frame containing the text data}

\item{id_col}{(string) the name of the column containing the unique id}

\item{data_col}{(string) the name of the column containing the text data}

\item{ngram_window}{(list) the minimum and maximum n-gram length, e.g. c(1,3)}

\item{stopwords}{(stopwords) the stopwords to remove, e.g. stopwords::stopwords("en", source = "snowball")}

\item{removalword}{(string) the word to remove}

\item{occ_rate}{(integer) the rate of occurence of a word to be removed}

\item{removal_mode}{(string) the mode of removal -> "most" or "least"}

\item{removal_rate_most}{(integer) the rate of most frequent words to be removed}

\item{removal_rate_least}{(integer) the rate of least frequent words to be removed}

\item{split}{(float) the proportion of the data to be used for training}

\item{seed}{(integer) the random seed for reproducibility}

\item{save_dir}{(string) the directory to save the results, default is "./results", if NULL, no results are saved}
}
\value{
the document term matrix
}
\description{
the function for creating a document term matrix
}
