% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1_topicsDtm.R
\name{topicsDtmEval}
\alias{topicsDtmEval}
\title{Summarize and Visualize your Document Term Matrix}
\usage{
topicsDtmEval(dtm)
}
\arguments{
\item{dtm}{(R_obj) The document term matrix -> output of topicsDtm function}
}
\value{
A named list containing:
\describe{
  \item{dtm_summary}{A dataframe of terms and their frequencies.}
  \item{frequency_plot}{A bar plot of all term frequencies with example terms.}
  \item{frequency_plot_30_least}{A bar plot of the 30 least frequent terms (if numer of terms > 30).}
  \item{frequency_plot_30_most}{A bar plot of the 30 most frequent terms (if numer of terms > 30).}
  \item{histogram_of_frequencies}{A histogram of term frequencies (this is the same information as
  in the frequency_plot but presented differently).}
}
}
\description{
This function creates a frequency table of your DTM and generates up to four plots for visualization
}
