% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{topicsPlot}
\alias{topicsPlot}
\title{The function to create lda wordclouds}
\usage{
topicsPlot(
  model,
  test,
  p_threshold = 0.05,
  grid = TRUE,
  dim = 2,
  color_scheme = "default",
  scale_size = FALSE,
  save_dir = "./results",
  figure_format = "png",
  seed = 42
)
}
\arguments{
\item{model}{(list) The trained model}

\item{test}{(list) The test results}

\item{p_threshold}{(integer) The p-value threshold to use for significance}

\item{grid}{(boolean) Set TRUE if plotting the topics grid}

\item{dim}{(numeric) Generate 1 dimensional color plots or 2 dimensioal color plots if grid = TRUE.}

\item{color_scheme}{(string 'default' or vector) The color scheme for plotted topic categories if grid = TRUE. The vector should contain 9 color codes.}

\item{scale_size}{(logical) Whether to scale the size of the words}

\item{save_dir}{(string) The directory to save the wordclouds}

\item{figure_format}{(string) Set the figure format, e.g., .svg, or .png.}

\item{seed}{(integer) The seed to set for reproducibility}
}
\value{
nothing is returned, the wordclouds are saved in the save_dir
}
\description{
The function to create lda wordclouds
}
