% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{topicsTest}
\alias{topicsTest}
\title{Test topics or n-grams}
\usage{
topicsTest(
  data,
  model = NULL,
  preds = NULL,
  ngrams = NULL,
  x_variable = NULL,
  y_variable = NULL,
  controls = c(),
  test_method = "linear_regression",
  p_adjust_method = "fdr",
  seed = 42,
  load_dir = NULL,
  save_dir
)
}
\arguments{
\item{data}{(tibble) The data containing the variables to be tested.}

\item{model}{(list) A trained model LDA-model from the topicsModel() function.}

\item{preds}{(tibble) The predictions from the topicsPred() function.}

\item{ngrams}{(list) output of the n-gram function}

\item{x_variable}{(string) The x variable name to be predicted, and to be plotted (only needed for regression or correlation)}

\item{y_variable}{(string) The y variable name to be predicted, and to be plotted (only needed for regression or correlation)}

\item{controls}{(vector) The control variables (not supported yet)}

\item{test_method}{(string) The test method to use, either "correlation","t-test", "linear_regression","logistic_regression", or "ridge_regression"}

\item{p_adjust_method}{(character) Method to adjust/correct p-values for multiple comparisons
(default = "none"; see also "holm", "hochberg", "hommel", "bonferroni", "BH", "BY",  "fdr").}

\item{seed}{(integer) The seed to set for reproducibility}

\item{load_dir}{(string) The directory to load the test from, if NULL, the test will not be loaded}

\item{save_dir}{(string) The directory to save the test, if NULL, the test will not be saved}

\item{group_var}{(string) The variable to group by (only needed for t-test)}
}
\value{
A list of the test results, test method, and prediction variable
}
\description{
Statistically test topics or n-grams in relation to one or two other variables using 
regression or t-test.
}
\examples{
\donttest{
# Test the topic document distribution in respect to a variable
save_dir_temp <- tempfile()

dtm <- topicsDtm(
  data = dep_wor_data$Depphrase, 
  save_dir =  save_dir_temp)

model <- topicsModel(
  dtm = dtm, # output of topicsDtm()
  num_topics = 20,
  num_top_words = 10,
  num_iterations = 1000,
  seed = 42,
  save_dir = save_dir_temp)
                     
preds <- topicsPreds(
 model = model, # output of topicsModel()
 data = dep_wor_data$Depphrase,
 save_dir = save_dir_temp)
                     
test <- topicsTest(
  model = model, # output of topicsModel()
  data=dep_wor_data,
  preds = preds, # output of topicsPreds()
  test_method = "linear_regression",
  x_variable = "Age",
  save_dir = save_dir_temp)
}                 
}
