% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{topicsWordclouds}
\alias{topicsWordclouds}
\title{The function to create lda wordclouds}
\usage{
topicsWordclouds(
  model,
  test,
  color_negative_cor = scale_color_gradient(low = "darkgreen", high = "green"),
  color_positive_cor = scale_color_gradient(low = "darkred", high = "red"),
  scale_size = FALSE,
  plot_topics_idx = NULL,
  p_threshold = 0.05,
  save_dir = "./results",
  figure_format = ".svg",
  seed = 42
)
}
\arguments{
\item{model}{(list) The trained model}

\item{test}{(list) The test results}

\item{color_negative_cor}{(R_obj) The color gradient for negative correlations}

\item{color_positive_cor}{(R_obj) The color gradient for positive correlations}

\item{scale_size}{(logical) Whether to scale the size of the words}

\item{plot_topics_idx}{(vector) The topics to plot determined by index}

\item{p_threshold}{(integer) The p-value threshold to use for significance}

\item{save_dir}{(string) The directory to save the wordclouds}

\item{figure_format}{(string) Set the figure format, e.g., .svg, or .png.}

\item{seed}{(integer) The seed to set for reproducibility}
}
\value{
nothing is returned, the wordclouds are saved in the save_dir
}
\description{
The function to create lda wordclouds
}
