% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{ldaPreds}
\alias{ldaPreds}
\title{The function to predict the topics of a new document with the trained model}
\usage{
ldaPreds(
  model,
  data,
  id_col,
  data_col,
  num_iterations = 100,
  seed = 42,
  save_dir = "./results",
  load_dir = NULL
)
}
\arguments{
\item{model}{(list) The trained model}

\item{data}{(tibble) The new data}

\item{id_col}{(string) The data column with the ids of the new documents}

\item{data_col}{(string) The data column with the text of the new documents}

\item{num_iterations}{(integer) The number of iterations to run the model}

\item{seed}{(integer) The seed to set for reproducibility}

\item{save_dir}{(string) The directory to save the model, if NULL, the predictions will not be saved}

\item{load_dir}{(string) The directory to load the model from, if NULL, the predictions will not be loaded}
}
\value{
A tibble of the predictions
}
\description{
The function to predict the topics of a new document with the trained model
}
