% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{topicsDtm}
\alias{topicsDtm}
\title{the function for creating a document term matrix}
\usage{
topicsDtm(
  data,
  ngram_window = c(1, 3),
  stopwords = stopwords::stopwords("en", source = "snowball"),
  removalword = "",
  occ_rate = 0,
  removal_mode = "none",
  removal_rate_most = 0,
  removal_rate_least = 0,
  split = 1,
  seed = 42L,
  save_dir = "./results",
  load_dir = NULL
)
}
\arguments{
\item{data}{(list) the list containing the text data with each entry belonging to a unique id}

\item{ngram_window}{(list) the minimum and maximum n-gram length, e.g. c(1,3)}

\item{stopwords}{(stopwords) the stopwords to remove, e.g. stopwords::stopwords("en", source = "snowball")}

\item{removalword}{(string) the word to remove}

\item{occ_rate}{(integer) the rate of occurence of a word to be removed}

\item{removal_mode}{(string) the mode of removal -> "none", "threshold", "absolute" or "percent", Threshold removes all words under a certain frequency or over a certain frequency as indicated by removal_rate_least and removal_rate_most, Absolute removes an absolute amount of terms that are most frequent and least frequent, Percent the amount of words indicated by removal_rate_least and removal_rate_most relative to the amount of terms in the matrix}

\item{removal_rate_most}{(integer) the rate of most frequent words to be removed, functionality depends on removal_mode}

\item{removal_rate_least}{(integer) the rate of least frequent words to be removed, functionality depends on removal_mode}

\item{split}{(float) the proportion of the data to be used for training}

\item{seed}{(integer) the random seed for reproducibility}

\item{save_dir}{(string) the directory to save the results, default is "./results", if NULL, no results are saved}

\item{load_dir}{(string) the directory to load from.}
}
\value{
the document term matrix
}
\description{
the function for creating a document term matrix
}
