% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2_2_topicsGrams.R
\name{topicsGrams}
\alias{topicsGrams}
\title{N-grams}
\usage{
topicsGrams(
  data,
  ngram_window = c(1, 3),
  stopwords = stopwords::stopwords("en", source = "snowball"),
  pmi_threshold = 0,
  top_frequent = 200
)
}
\arguments{
\item{data}{(tibble) The data}

\item{ngram_window}{(list) the minimum and maximum n-gram length, e.g. c(1,3)}

\item{stopwords}{(stopwords) the stopwords to remove, e.g. stopwords::stopwords("en", source = "snowball")}

\item{pmi_threshold}{(integer) The pmi threshold, if it shall not be used set to 0}

\item{top_frequent}{(integer) The number of most frequently occuring ngrams to included in the output.}
}
\value{
A list containing tibble of the ngrams with the frequency and probability and 
a tibble containing the relative frequency of the ngrams for each user
}
\description{
The function computes ngrams from a text
}
