% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{topicsGrams}
\alias{topicsGrams}
\title{N-grams}
\usage{
topicsGrams(
  data,
  ngram_window = c(1, 3),
  stopwords = stopwords::stopwords("en", source = "snowball"),
  top_n = NULL,
  pmi_threshold = 0
)
}
\arguments{
\item{data}{(tibble) The data}

\item{ngram_window}{(list) the minimum and maximum n-gram length, e.g. c(1,3)}

\item{stopwords}{(stopwords) the stopwords to remove, e.g. stopwords::stopwords("en", source = "snowball")}

\item{top_n}{(integer) The number of most occuring ngrams to included in the output for every ngram type}

\item{pmi_threshold}{(integer) The pmi threshold, if it shall not be used set to 0}
}
\value{
A list containing tibble of the ngrams with the frequency and probability and 
a tibble containing the relative frequency of the ngrams for each user
}
\description{
The function computes ngrams from a text
}
