% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{topicsGrams}
\alias{topicsGrams}
\title{N-grams}
\usage{
topicsGrams(data, n = 2, sep = " ", top_n = NULL, pmi_threshold = 0)
}
\arguments{
\item{data}{(tibble) The data}

\item{n}{(integer) The length of ngram}

\item{sep}{(string) The separator}

\item{top_n}{(integer) The number of top ngrams to be displayed}

\item{pmi_threshold}{(integer) The pmi threshold, if it shall not be used set to 0}
}
\value{
A list containing tibble of the ngrams with the frequency and probability and 
a tibble containing the relative frequency of the ngrams for each user
}
\description{
The function computes ngrams from a text
}
