% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{topicsPlot}
\alias{topicsPlot}
\title{The function to create lda wordclouds}
\usage{
topicsPlot(
  model,
  test,
  p_threshold = 0.05,
  grid_plot = TRUE,
  dim = 2,
  color_scheme = "default",
  scatter_legend_way_popout_topics = c("mean", "max_x", "max_y"),
  scatter_legend_user_spec_topics = NULL,
  scatter_legend_topic_num = FALSE,
  scale_size = FALSE,
  save_dir = "./results",
  figure_format = "svg",
  width = 10,
  height = 8,
  max_size = 10,
  seed = 42,
  scatter_legend_popout_dot_size = 15,
  scatter_legend_bg_dot_size = 9,
  grid_legend_title = "legend_title",
  grid_legend_title_size = 5,
  grid_legend_title_color = "black",
  grid_legend_x_axes_label = "legend_x_axes_label",
  grid_legend_y_axes_label = "legend_y_axes_label",
  grid_legend_number_color = "black",
  grid_legend_number_size = 5
)
}
\arguments{
\item{model}{(list) The trained model}

\item{test}{(list) The test results}

\item{p_threshold}{(integer) The p-value threshold to use for significance}

\item{grid_plot}{(boolean) Set TRUE if plotting the topics grid}

\item{dim}{(numeric) Generate 1 dimensional color plots or 2 dimensioal color plots if grid = TRUE}

\item{color_scheme}{(string 'default' or vector) The color scheme for plotted topic categories if grid = TRUE. The vector should contain 9 color codes.}

\item{scatter_legend_way_popout_topics}{(string) The way to filter topics to popout in the scatter legend. Can be either "mean", "max_x", or "max_y"}

\item{scatter_legend_user_spec_topics}{(vector) User can specify which topic to be poped out in the scatter legend. Should be like c("t_1", "t_2", ...). If set, way_popout_topics will have no effect.}

\item{scatter_legend_topic_num}{(boolean) Allow showing the topic number or not in the scatter legend}

\item{scale_size}{(logical) Whether to scale the size of the words}

\item{save_dir}{(string) The directory to save the wordclouds}

\item{figure_format}{(string) Set the figure format, e.g., .svg, or .png.}

\item{width}{(integer) The width of the topic (units = "in").}

\item{height}{(integer) The width of the topic (units = "in").}

\item{max_size}{(integer) The max size of the words.}

\item{seed}{(integer) The seed to set for reproducibility}

\item{scatter_legend_popout_dot_size}{(integer) The dot size of scatter legend to popout}

\item{scatter_legend_bg_dot_size}{(integer) The dot size of scatter legend in the background}

\item{grid_legend_title}{The title of grid topic plot if grid_plot = TRUE}

\item{grid_legend_title_size}{The size of the title of the plot if grid_plot = TRUE}

\item{grid_legend_title_color}{The color of the legend title if grid_plot = TRUE}

\item{grid_legend_x_axes_label}{The label of the x axes if grid_plot = TRUE}

\item{grid_legend_y_axes_label}{The label of the y axes if grid_plot = TRUE}

\item{grid_legend_number_color}{The color in the text in the legend if grid_plot = TRUE}

\item{grid_legend_number_size}{The color in the text in the legend if grid_plot = TRUE}
}
\value{
nothing is returned, the wordclouds are saved in the save_dir
}
\description{
The function to create lda wordclouds
}
