% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/4_1_topicsPlot.R
\name{topicsPlotOverview}
\alias{topicsPlotOverview}
\title{Combine topics and distribution legend (experimental)}
\usage{
topicsPlotOverview(plot_list, overview_plot_type, title_font = "sans")
}
\arguments{
\item{plot_list}{(list) Output from the topicsPlot function.}

\item{overview_plot_type}{(character) Number of dimensions to plot (1 or 2).}
}
\value{
An overview plot including topics and distribution legend
}
\description{
Combine topics and distribution legend (experimental)
}
